# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brain Bridge API Documentation",
        "description": "API documentation for Brain Bridge application",
        "contact": {
            "email": "gmzesan7767@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get list of all subjects",
                "description": "Returns all subjects from the database",
                "operationId": "getSubjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Create a new subject",
                "description": "Creates a new subject",
                "operationId": "storeSubject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mathematics"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subject created successfully"
                    }
                }
            }
        },
        "/api/subjects/{id}": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get subject information",
                "description": "Returns subject data",
                "operationId": "getSubjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subject id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Update an existing subject",
                "operationId": "updateSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subject ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Physics"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subject updated successfully"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Delete a subject",
                "operationId": "deleteSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subject ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject deleted successfully"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                }
            }
        },
...
    },
...
}